flask_app/requirements.txt
flask
psycopg2-binary
celery
redis